Add DOCS-WIP
We've needed a way to queue/note/sync docs changes with our manuals, migration
guides and release notes. Add a file which we can use to "stage" information.
We can then ask users submitting changes which need docs changes to add
information here. This will then make it easier for the docs to be brought
into sync as we'll have a much better idea of when something merged and
what changes are needed.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>